Refactor continuation turn and local runner seams
